+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v3.0.0
+
+### Added
+* `DrawerModal` component (#83)
+* `Pill` component (#84)
+
+### Breaking Change
+* Deprecated `fonts` object. Introduces the `styled-components` theme. Fonts are now configured inside the theme
+```tsx
+<ThemeProvider theme={{ fonts: { heading, body, serif, sans } }}>
+  ...
+</ThemeProvider>
+```
+
 ## v2.7.0
 
 ### Added